requirements.txt
attrs==20.3.0
Automat==20.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
oauth2client==4.1.3
packaging==20.9
parsel==1.6.0
Protego==0.1.16
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyDrive==1.3.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
queuelib==1.5.0
requests==2.25.1
rsa==4.7.2
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
Twisted==21.2.0
uritemplate==3.0.1
urllib3==1.26.4
w3lib==1.22.0
zope.interface==5.3.0
